Fix type of `OrderResponseItem.Quantity` (`int` -> `decimal`) (#8)

+++ openTRANS/ORDERRESPONSE/OrderResponseItem.cs
@@ -21,7 +21,7 @@ namespace openTRANS
         public ProductComponents ProductComponents = new ();
 
         [XmlElement("QUANTITY")]
-        public int Quantity;
+        public decimal Quantity;
 
         [XmlElement("ORDER_UNIT", Namespace = Namespace.bmecat)]
         public string OrderUnit;
